+++ packages/calcite-components/src/components/dialog/dialog.e2e.ts
@@ -1,3 +1,4 @@
+// @ts-strict-ignore
 import { newE2EPage, E2EPage } from "@arcgis/lumina-compiler/puppeteerTesting";
 import { describe, expect, it, vi } from "vitest";
 import {
@@ -1051,8 +1052,8 @@ describe("calcite-dialog", () => {
       let computedStyle = await container.getComputedStyle();
       const initialBlockSize = computedStyle.blockSize;
       const initialInlineSize = computedStyle.inlineSize;
-      const initialHeight = parseInt(initialBlockSize, 10);
-      const initialWidth = parseInt(initialInlineSize, 10);
+      const initialHeight = parseInt(initialBlockSize);
+      const initialWidth = parseInt(initialInlineSize);
 
       await dispatchDialogKeydown({ page, key: "ArrowUp", shiftKey: true });