+++ packages/calcite-components/src/components/input/common/input.ts
@@ -1,3 +1,4 @@
+// @ts-strict-ignore
 export type InputComponent = NumericInputComponent | TextualInputComponent | DateTimeInputComponent;
 
 export interface DateTimeInputComponent {